Fix mark as selected for array of objects
Fixing the bug: for array of objects only one field gets marked as selected, because this.optionObjectComparator owerwrites the results of previous checks. This happens when more then 2 items are selected and we deal with array of objects. Then, when checking n's selected option, n-1 is always marked as not selected.

This simple fix corrects the situation.