fix(Paginated): fix bugs in Pagination use case example in docs
Reset the current offset each type the search query changes; otherwise
there may be no results shown if the user clicked forward one or more
pages and then typed to search.

Replace instances of magic number 10 with the "limit" variable;
otherwise changing the value of the "limit" data prop will result in
buggy behavior because clicking forward and back still updates the
offset by 10.

Convert query and country name strings to lower case before filtering
the list so that the user doesn't have to match the exact case of the
country names.